Added PRINT #-3, feature
Using LOCATE x,y you can set the column and horizontal position where to print text on the PMODE 4 screen.  The x value must be between 0 and 31, the y value can be from 0 to 184.  You can also select the font from the command line with the new -fFont option where the Font name is either Arcade or CoCoT1
 Use PRINT #-3,"Hello World".